requirements.txt
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.25.2
tqdm==4.66.1
python-dateutil==2.8.2
pydantic==2.5.0
loguru==0.7.2
